Added first version of reddit reader.
The small project makes use of the http package in order to send
a get request to http://reddit.com/r/golang.json. This url seems to
serve a json data blob which includes the titles of currently discussed
go topics.
The program that just prints the titles one after the other.